refactor(interpreter): split statement scheduling and activation tests
